Split up main.ts, fix some code issues
