+++ src/main.ts
@@ -1,15 +1,6 @@
-import { App, Editor, MarkdownView, Notice, Plugin, PluginSettingTab, Setting } from 'obsidian';
-import { transferNote } from 'transfer';
-
-export interface VaultTransferSettings {
-	outputVault: string;
-	outputFolder: string;
-}
-
-const DEFAULT_SETTINGS: VaultTransferSettings = {
-	outputVault: '',
-	outputFolder: '',
-}
+import { addCommands } from 'commands';
+import { Plugin } from 'obsidian';
+import { DEFAULT_SETTINGS, SettingTab, VaultTransferSettings } from 'settings';
 
 export default class VaultTransferPlugin extends Plugin {
 	settings: VaultTransferSettings;
@@ -17,17 +8,8 @@ export default class VaultTransferPlugin extends Plugin {
 	async onload() {
 		await this.loadSettings();
 
-		// This adds an editor command that can perform some operation on the current editor instance
-		this.addCommand({
-			id: 'transfer-note-to-vault',
-			name: 'Transfer current note to other vault',
-			editorCallback: (editor: Editor, view: MarkdownView) => {
-				this.app;
-				transferNote(editor, view, this.app, this.settings);
-			}
-		});
+		addCommands(this);
 
-		// This adds a settings tab so the user can configure various aspects of the plugin
 		this.addSettingTab(new SettingTab(this.app, this));
 	}
 
@@ -37,44 +19,5 @@ export default class VaultTransferPlugin extends Plugin {
 
 	async saveSettings() {
 		await this.saveData(this.settings);
-	}
-}
-
-class SettingTab extends PluginSettingTab {
-	plugin: VaultTransferPlugin;
-
-	constructor(app: App, plugin: VaultTransferPlugin) {
-		super(app, plugin);
-		this.plugin = plugin;
-	}
-
-	display(): void {
-		const { containerEl } = this;
-
-		containerEl.empty();
-
-		containerEl.createEl('h2', { text: 'Settings' });
-
-		new Setting(containerEl)
-			.setName('Output Vault')
-			.setDesc('The full file path to the other vault root folder.')
-			.addText(text => text
-				.setPlaceholder('D:/MyVault')
-				.setValue(this.plugin.settings.outputVault)
-				.onChange(async (value) => {
-					this.plugin.settings.outputVault = value;
-					await this.plugin.saveSettings();
-				}));
-
-		new Setting(containerEl)
-			.setName('Output Folder')
-			.setDesc('The folder within the vault the file should be copied to.')
-			.addText(text => text
-				.setPlaceholder('Unsorted/Transfer')
-				.setValue(this.plugin.settings.outputFolder)
-				.onChange(async (value) => {
-					this.plugin.settings.outputFolder = value;
-					await this.plugin.saveSettings();
-				}));
 	}
 }
